Changed fahrenheit to use floating point numbers.

+++ fahrenheit.c
@@ -4,11 +4,11 @@
 
 #include <stdio.h>
 /* print Fahrenheit-Celsius table
- * for fahr = 0, 20, ..., 300 */
+ * for fahr = 0, 20, ..., 300; floating-point version */
 
 main()
 {
-    int fahr, celsius;
+    float fahr, celsius;
     int lower, upper, step;
 
     lower = 0;      /* lower limit of temperature table */
@@ -17,8 +17,8 @@ main()
 
     fahr = lower;
     while (fahr <= upper) {
-        celsius = (5 / 9) * (fahr-32);
-        printf("%3d %6d\n", fahr, celsius);
+        celsius = (5.0 / 9.0) * (fahr - 32);
+        printf("%3.0f %6.1f\n", fahr, celsius);
         fahr = fahr + step;
     }
 }
